警告の修正 サブウェポンの実装 マップの修正 テストシーンの追加 アイテムクラスの追加

+++ Team16/Actor/Bulletes/ChangeBullet.cpp
@@ -35,7 +35,7 @@ void ChangeBullet::update(float deltaTime)
 
 void ChangeBullet::draw(Renderer * renderer, Renderer3D* renderer3D)
 {
-	DrawCircle(b_mPosittion.x + 64 / 2, b_mPosittion.y + 32, b_mCircleSize, GetColor(0, 0, 255), TRUE);
+	DrawCircle((int)(b_mPosittion.x + 64 / 2), (int)(b_mPosittion.y + 32), (int)b_mCircleSize, GetColor(0, 0, 255), TRUE);
 }
 
 void ChangeBullet::hit(BaseObject & other)
