+++ Team16/Actor/Bulletes/LaserBullet.cpp
@@ -80,7 +80,7 @@ void LaserBullet::draw(Renderer * renderer, Renderer3D * renderer3D)
 	float angle = atan2(a.y, a.x)* 180.0f / DX_PI_F;
 
 	//renderer->draw2D("bullet1", Vector2(b_mPosittion.x, b_mPosittion.y), Vector2(0, 0), Vector2(12, 16), Vector2(6, 8), Vector2(1.0f, 1.0f), angle - 90.0f, 255);
-	DrawCircle(b_mPosittion.x + 64.0f / 2, b_mPosittion.y + 64.0f / 2, b_mCircleSize, GetColor(255, 255, 0), TRUE);
+	DrawCircle((int)(b_mPosittion.x + 64.0f / 2), (int)(b_mPosittion.y + 64.0f / 2), (int)b_mCircleSize, GetColor(255, 255, 0), TRUE);
 
 	if (b_mType == Type::ENEMY_BULLET)
 	{
@@ -107,14 +107,14 @@ void LaserBullet::hit(BaseObject & other)
 		b_mIsDeath = true;
 	}
 
-	DrawCircle(b_mPosittion.x + 64.0f / 2, b_mPosittion.y + 64.0f / 2, b_mCircleSize, GetColor(255, 255, 0), TRUE);
+	DrawCircle((int)(b_mPosittion.x + 64.0f / 2), (int)(b_mPosittion.y + 64.0f / 2), (int)b_mCircleSize, GetColor(255, 255, 0), TRUE);
 }
 
 Vector2 LaserBullet::RotationZ(float ang)
 {
 	Vector2 v;
 	ang = ang + 45.0f;
-	ang = ang * PI / 180.0;
+	ang = ang * PI / 180.0f;
 
 	float sin = sinf(ang);
 	float cos = cosf(ang);
